Add validation tools, an sqlite3 database, and fix data
Objective:
The primary objective of this commit is to provide an SQL interface
querying the Starfield data. This enables users to take advantage
of SQL to either directly query the data for what they're looking
for, or to build apps/tooling on top of this work for other purposes.

Details:
This commit adds 2 scripts (create_db.py, restructure.py).

restructure.py reads the starfield_data_updated.json file, validates it,
and then stores the data in a flattened format in galaxy.json.

create_db.py takes the galaxy.json output of restructure.py and
inserts the data into a sqlite3 database file.